+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ---
 
-- Fixed HTML intellisense (html.json was missed in the package)
+- **HTML**. Fixed intellisense (html.json was missed in the package)
+- **Styles**. Textbased controls. Added `opacity: 1` for `prefix/postfix` as fix when controls in a row and it resizes on hover
 
 ---