+++ CHANGELOG.md
@@ -24,7 +24,7 @@
 ---
 
 - [WUPTimeObject](src/objects/timeObject.ts). Add init by date to constructor
-- [DateControl](src/controls/date.ts). Add sync with time (value, min, max)
+- [DateControl](src/controls/date.ts). Add sync with [TimeControl](src/controls/time.ts) (value, min, max)
 
 ---